[WIP] Handle case for resolving dependencies when replacement packages exist
Add a case which picks the sole package that isn't an alternate package.

Things which depend on alternate packages will need to explicitly specify
the alternate packages it requires.

e.g. Tegra grate-driver provides libdrm-grate which is an alternate libdrm.
The package device-samsung-p4wifi explicity depends on libdrm-grate.
Other things which don't depend on the libdrm alternate will to work as
expected without accidentally having libdrm-grate installed.

This will also resolve issue #1380 where the incorrect GTK+2 package
would be installed.


TODO:
There still is an uncertain case where multiple packages provide a
dependency and they are not replacements for another provider.
What do if there is no singular original provider which isn't a
replacement.
Should pmbootstrap handle this some how?
Should pmbootstrap throw an error and force the user to resolve the
conflict?
This could result in more unexpected runtime behavior if one packaged is
picked but the user expected another.